requirements.txt
asgiref==3.10.0
Django==5.2.7
pillow==12.0.0
sqlparse==0.5.3
djangorestframework==3.15.0
django-cors-headers==4.5.0
